refactor(api): merge and rework tests/shared/unit/domain/errors_test.js
